+++ NEWS.md
@@ -1,3 +1,18 @@
+# BiometryTraining 0.9.0
+
+# Major changes
+
+- `resplot()` function now supports models from the packages `lme4`, `nlme`, and `sommer`.
+- `mct.out()` function now supports models from the packages `lme4` and `nlme`.
+
+## Minor changes
+
+- Fixed some bugs in `logl.test()`.
+- Unknown arguments in `design()` are no longer silently ignored (#77)
+- Enabled a way for split plot designs to be arranged row-wise or column-wise (#68)
+- Added split plot treatment labels to the output plot (#78)
+- Prevented misspelled arguments passed in the ellipsis from being silently ignored (#77)
+
 # BiometryTraining 0.8.0
 
 ## Minor changes